timer: Implement skeleton timer device
The Timer class represents a Timer hardware device 
and as such it can not be copied or moved. It has
a thread of execution for itself - it sleeps for a
period of time configured by the timer_cfg reg.

The timer will generate interrupts to the CPU on
every timeout - the timeout happens every time the
timer thread wakes up.

The Timer is a RAII class - it's thread starts upon
construction - and is joined upon destruction. The
destruction might generate a delay in program
execution based on the current configured value
in the timer_cfg register.