terminal: Implement skeleton terminal device
The Terminal class represents a Terminal hardware
device and as such it can not be copied or moved. It
has a thread of execution for itself - it reads user
input and keeps the last read byte in it's data_in
register (and will raise an interrupt to the CPU when
it has read user input). When writing to the data_out
register - Terminal prints the char written to the
output.

The Terminal is a RAII class - it's thread starts upon
construction - and is joined upon destruction.
The destruction might generate a wait for user input
on Win32 as Terminal blocks on user input.