memory: Implement skeleton memory
The Memory class represents and abstracts
the hardware memory of the emulator. All
memory accesses should go through the
interface of this class. This interfaces also
supports memory mapped registers of the
CPU and other devices in the emulator.

TODO: this is a skeleton implementation with
no support for real memory mapped accesses
to registers