cpu: Implement structure and interrupt interface
The Cpu class represents a physical CPU of the
emulator. The structure contains the registers of
the CPU - GP regs, memory mapped regs and the
PSW along with all necessary definitions.

The instructions structs represent an abstraction
for CPU instructions to be executed - implemented
along with all the necessary definitions regarding
opcodes, addressing modes, descriptors, etc.

The interrupt interface is simple - just one skeleton
function, along with interrupts constants definitions.
We need this for hardware interrupts as well as
software interrupts in the emulation.