assembler: Refactor .h file extensions to .hpp
This is done to avoid conflicts with standard library
filenames. The quote include already does this,
as does the first capital letter on Linux systems,
but intellisense is confused by the same filenames,
and this ensures name clashing with standard
file names, as no standard header has
the .hpp extension.